Add templating for <head>, header, & footer
This will allow a header and footer file to be inserted on all pages
making it easy to manage the header and footer in one location.
This will also populate some <head> content such as links and other
content that will be the same on all pages.
Also added Jquery while figuring this out; leaving it in since it will
probably be used later.