Weight loss info, pictures
I added the missing information that belonged on the weight loss page
and added the pictures that will go on that page and added a better
picture of the focus attention and rescue gel as well as the
chlorophyll bottle. I also added a pic of my logo to go at the top of
each page under the banner to the left.