Updated gutter image
There was a few more too cut out that I thought were good, so I cut
out. It helped fix the duplicates and scarcity at the bottom of the
gutter. There are two pics that are a bit out of focus and if I have
time I will re-take photos of them. Parsley and purple pansies, for now
I am not worrying about them.